docs(776): the Gate 1 flash tutorial, and the model it needs to import
The Gate 1 fixture landed as a benchmark harness, which is the wrong home
for the half of it that is a *model*. `benchmarks/` does not ship in the
wheel, so nothing a user installs could import the flash, and a tutorial
that reached into a benchmark directory would be off-pattern next to the
other three application cases.

So the model, its complementarity pairs, the three lowerings and the
independent oracle move to `pounce.examples.flash_mpcc`, alongside
`phase_envelope` and `catalyst_pellet`. `benchmarks/flash/` keeps what it
should have kept in the first place -- routes, the measurement protocol,
the traversal, provenance and the report -- and imports the model like any
other caller. Keeping the two apart is what stops the harness from
becoming a second, subtly different copy of the physics, which is the
failure mode the whole cross-check exists to catch.

Notebook 38 is the narrative walk-through, following the template the
epic asks for: motivation, equations, baseline, formulation, diagnostics,
verification, result, limitations. Five figures, executed and saved:

* the vapor fraction across the path, with both switch points marked;
* **the two stability slacks** -- the plot that shows the complementarity
  structure directly, each slack leaving zero at its own switch point and
  both vanishing between them;
* agreement with the independent reference against Gate 0's `sqrt(tol)`
  complementarity floor, eight orders below it, peaking at the two
  biactive points where it should;
* the same solve read as a T-x-y diagram, where the branches continue
  outside the two-phase lens because the incipient phase is carried as
  real variables at every temperature;
* the reverse-mode Hessian error: fourteen orders at the bubble point,
  machine precision everywhere else.

The tutorial does not overclaim, and the measurement is why. Ordinary
successive substitution solves every point on this path correctly in 7-10
sweeps against roughly a hundred interior-point iterations, so the
notebook says in section 2 that the complementarity form *loses* as a way
to compute one flash. The case for it is composability -- a procedure with
an inner loop, a regime branch and a clipped Rachford-Rice endpoint whose
derivative is identically zero in the single-phase regions cannot be one
block of a simultaneous transcription -- and the notebook's job is to
establish correctness where the answer is independently known, before the
formulation is used where it is not.

Two cells reproduce findings rather than assert them: the defective
isofugacity row is evaluated beside the correct one and shown to be
exactly zero in the two-phase region and `ln(Sy/Sx)` outside it; and the
`prod_eq` finishing solve is run from the continuation's own answer and
reports `Not_Enough_Degrees_Of_Freedom` with the six equality rows named
against five variables.

`docs/src/flash-mpcc.md` is the user-facing page, with a SUMMARY entry
beside the other application cases.
